Create student grade tracker
The program prompts the user to input the number of students and then iterates through a loop to collect each student's name and score. The information is stored in the array of Student objects. Finally, the program prints out each student's name and calculates grades by utilizing the getters from the Student class.